Migrate auth/session related methods from index.js to AppContext - #62 sennetconsortium/portal-ui#771
