+++ src/pages/index.js
@@ -4,13 +4,12 @@
 import DataTable from "../components/DataTable";
 import Login from "../components/Login";
 import Blank from "../components/Blank";
-import Image from 'next/image';
-import { useState, useEffect } from "react";
-import {ingest_api_users_groups} from "../service/ingest_api";
-import process from "next/dist/build/webpack/loaders/resolve-url-loader/lib/postcss";
+import { useState, useEffect, useContext } from "react";
+import AppContext from "../context/AppContext";
 
 
 function App({ entity_type, upload_id, page, page_size, sort_field, sort_order, filters }) {
+    const {globusInfo, globusToken, handleLogin, handleLogout, checkLocals, isLoading, isAuthenticated, unauthorized} = useContext(AppContext)
     const [entityType, setEntityType] = useState(entity_type);
     const [selectUploadId, setSelectUploadId] = useState(upload_id);
     const [initialPage, setInitialPage] = useState(page);
@@ -18,112 +17,10 @@ function App({ entity_type, upload_id, page, page_size, sort_field, sort_order, 
     const [sortField, setSortField] = useState(sort_field);
     const [sortOrder, setSortOrder] = useState(sort_order);
     const [tableFilters, setTableFilters] = useState(filters);
-    const [isLoading, setIsLoading] = useState(true);
-    const [isAuthenticated, setIsAuthenticated] = useState(false);
-    const [globusInfo, setGlobusInfo] = useState(null);
-    const [globusToken, setGlobusToken] = useState(null);
-    const [unauthorized, setUnauthorized] = useState(false);
-
 
     if (initialPage === undefined) {
         setInitialPage(1);
     }
-
-    const checkLocals = (authenticated, initialInfo) => {
-        if (initialInfo) {
-            setGlobusInfo(initialInfo);
-            setGlobusToken(JSON.parse(initialInfo).groups_token)
-            checkToken(JSON.parse(initialInfo).groups_token).then(validate => {
-                setIsAuthenticated(authenticated && validate.hubmapUser);
-                if (validate.hubmapUser === false && validate.invalidToken === false) {
-                    setUnauthorized(true);
-                }
-                setIsLoading(false);
-            })
-            .catch(error => {
-                console.log(error);
-                setIsLoading(false);
-            })
-
-        } else {
-            setIsLoading(false);
-        }
-    }
-
-    const loginUrl = `${process.env.NEXT_PUBLIC_APP_BACKEND_URL}/data-ingest-board-login`
-    const logoutUrl = `${process.env.NEXT_PUBLIC_APP_BACKEND_URL}/data-ingest-board-logout`
-
-    const handleLogin = () => {
-        window.location.href = loginUrl
-    };
-
-    const handleLogout = () => {
-        window.location.href = logoutUrl
-        setGlobusToken(null);
-        setGlobusInfo(null);
-        setUnauthorized(false);
-        localStorage.removeItem("info");
-        localStorage.removeItem("isAuthenticated");
-        setIsAuthenticated(false);
-    }
-
-    const checkToken = (tokenInfo) => {
-        let hubmapUser = false;
-        let invalidToken = false;
-
-        return new Promise((resolve, reject) => {
-            try {
-                ingest_api_users_groups(tokenInfo).then((results) => {
-                    if (results && results.status === 200) {
-                        hubmapUser = results.results.some(obj => obj.displayname === "HuBMAP Read");
-                    }
-                    if (results && results.status === 401) {
-                        invalidToken = true;
-                    }
-                    resolve({hubmapUser, invalidToken});
-                }).catch(error => {
-                    console.log(error);
-                    reject(error);
-                });
-            } catch(error) {
-                console.log(error)
-                reject(error);
-            }
-        });
-    }
-
-    useEffect(() => {
-        let url = new URL(window.location.href);
-        let info = url.searchParams.get("info");
-        if (info) {
-            window.history.pushState(null, null, `/`);
-            localStorage.setItem("info", info);
-            localStorage.setItem("isAuthenticated", "true");
-            setGlobusInfo(info);
-            setGlobusToken(JSON.parse(info).groups_token);
-            checkToken(JSON.parse(info).groups_token).then(tokenValidation => {
-                // if (tokenValidation.hubmapUser) {
-                //     localStorage.setItem("isAuthenticated", "true");
-                //     setIsAuthenticated(true);
-                // } else {
-                //     setIsAuthenticated(false);
-                //     setUnauthorized(true);
-                // }
-                if (tokenValidation.hubmapUser) {
-                    localStorage.setItem("isAuthenticated", "true");
-                    setIsAuthenticated(true);
-                } else {
-                    if (tokenValidation.invalidToken === false) {
-                        setUnauthorized(true);
-                    }
-                }
-
-            }).catch(error => {
-                console.log(error)
-            })
-            setIsLoading(false);
-        }
-    }, [globusToken, globusInfo, isAuthenticated, isLoading])
 
     return (
         <div className="App">
